refactor(conversion): extract tabular conversion pipeline
Move CSV/GeoJSON/Parquet conversion and shared schema helpers into udata_hydra.conversion. Use tempfile.gettempdir() when resolving queued analysis paths so re-open matches download_resource. Restore csv_to_parquet patch target in tests after the import move.
