+++ udata_hydra/analysis/csv.py
@@ -1,40 +1,19 @@
-import csv as stdcsv
 import hashlib
 import json
 import logging
 import os
-import sys
 from datetime import datetime, timezone
 
 from asyncpg import Record
-from asyncpg.exceptions import UndefinedTableError
 from csv_detective import routine as csv_detective_routine
 from csv_detective import validate_then_detect
-from csv_detective.detection.engine import engine_to_file
-from csv_detective.utils import sanitize_for_json
-from progressist import ProgressBar
-from slugify import slugify
-from sqlalchemy import (
-    JSON,
-    BigInteger,
-    Boolean,
-    Column,
-    Date,
-    DateTime,
-    Float,
-    Integer,
-    LargeBinary,
-    MetaData,
-    String,
-    Table,
-)
-from sqlalchemy.dialects.postgresql import asyncpg
-from sqlalchemy.schema import CreateIndex, CreateTable, Index
 
-from udata_hydra import config, context
+from udata_hydra import config
 from udata_hydra.analysis import helpers
-from udata_hydra.analysis.geojson import csv_to_geojson_and_pmtiles
-from udata_hydra.db import compute_insert_query, db_col_name
+from udata_hydra.analysis.tables_index import csv_to_db_index, get_previous_analysis
+from udata_hydra.conversion.csv_to_db import csv_to_db
+from udata_hydra.conversion.csv_to_geojson_and_pmtiles import csv_to_geojson_and_pmtiles
+from udata_hydra.conversion.csv_to_parquet import csv_to_parquet
 from udata_hydra.db.check import Check
 from udata_hydra.db.resource import Resource
 from udata_hydra.db.resource_exception import ResourceException
@@ -46,35 +25,11 @@ from udata_hydra.utils import (
     handle_parse_exception,
     remove_remainders,
 )
-from udata_hydra.utils.casting import generate_records
-from udata_hydra.utils.minio import MinIOClient
-from udata_hydra.utils.parquet import db_to_parquet, save_as_parquet
+
+# Re-exported for backwards-compatible imports (e.g. udata_hydra/cli.py).
+__all__ = ["analyse_csv", "csv_detective_routine"]
 
 log = logging.getLogger("udata-hydra")
-
-# Increase CSV field size limit to maximum possible
-# https://stackoverflow.com/a/15063941
-field_size_limit = sys.maxsize
-while True:
-    try:
-        stdcsv.field_size_limit(field_size_limit)
-        break
-    except OverflowError:
-        field_size_limit = int(field_size_limit / 10)
-
-PYTHON_TYPE_TO_PG = {
-    "string": String,
-    "float": Float,
-    "int": BigInteger,
-    "bool": Boolean,
-    "json": JSON,
-    "date": Date,
-    "datetime": DateTime,
-    "datetime_aware": DateTime(timezone=True),
-    "binary": LargeBinary,
-}
-
-minio_client = MinIOClient(bucket=config.MINIO_PARQUET_BUCKET, folder=config.MINIO_PARQUET_FOLDER)
 
 
 async def analyse_csv(
@@ -218,261 +173,3 @@ async def analyse_csv(
 
         # Reset resource status to None
         await Resource.update(resource_id, {"status": None})
-
-
-async def get_previous_analysis(resource_id: str) -> dict | None:
-    db = await context.pool("csv")
-    q = (
-        "SELECT parsing_table, csv_detective FROM tables_index "
-        f"WHERE resource_id='{resource_id}' ORDER BY created_at DESC LIMIT 1"
-    )
-    res = await db.fetch(q)
-    if not res:
-        return None
-    analysis = json.loads(res[0]["csv_detective"])
-    # the csv_detective column is JSONB, so keys are reordered compared to the actual table
-    # so we get the right order from the table
-    try:
-        rows: list[Record] = await db.fetch(f'SELECT * FROM "{res[0]["parsing_table"]}" LIMIT 1')
-    except UndefinedTableError:
-        log.error(f"Could not find table {res[0]['parsing_table']} for resource {resource_id}")
-        return None
-    if len(rows) == 0:
-        return None
-    # the __id column is generated by hydra, not natively in the data
-    table_columns = [col for col in rows[0].keys() if col != "__id"]
-    analysis["columns"] = {col: analysis["columns"][col] for col in table_columns}
-    return analysis
-
-
-def compute_create_table_query(
-    table_name: str, columns: dict, indexes: dict[str, str] | None = None
-) -> str:
-    """Use sqlalchemy to build a CREATE TABLE statement that should not be vulnerable to injections"""
-
-    metadata = MetaData()
-    table = Table(table_name, metadata, Column("__id", Integer, primary_key=True))
-
-    for col_name, col_type in columns.items():
-        table.append_column(Column(col_name, PYTHON_TYPE_TO_PG.get(col_type, String)))
-
-    indexes_labels = {}
-    if indexes:
-        for col_name, index_type in indexes.items():
-            if index_type not in config.SQL_INDEXES_TYPES_SUPPORTED:
-                log.error(
-                    f'Index type "{index_type}" is unknown or not supported yet! '
-                    f"Index for column {col_name} was not created."
-                )
-                continue
-
-            else:
-                if index_type == "index":
-                    index_name = f"{table_name}_{slugify(col_name)}_idx"
-                    indexes_labels[index_name] = {"column": col_name, "type": index_type}
-                    try:
-                        table.append_constraint(Index(index_name, col_name))
-                    except KeyError:
-                        raise KeyError(
-                            f'Error creating index "{index_name}" on column "{col_name}". '
-                            f'Does the column "{col_name}" exist in the table?'
-                        )
-                # TODO: other index types. Not easy with sqlalchemy, maybe use raw sql?
-
-    compiled_query = CreateTable(table).compile(dialect=asyncpg.dialect())
-    query: str = compiled_query.string
-
-    # Add the index creation queries to the main query
-    for index in table.indexes:
-        log.debug(
-            f'Creating {indexes_labels[index.name]["type"]} on column "{indexes_labels[index.name]["column"]}"'
-        )
-        query_idx = CreateIndex(index).compile(dialect=asyncpg.dialect())
-        query: str = query + ";" + query_idx.string
-
-    # compiled query will want to write "%% mon pourcent" VARCHAR but will fail when querying "% mon pourcent"
-    # also, "% mon pourcent" works well in pg as a column
-    # TODO: dirty hack, maybe find an alternative
-    query = query.replace("%%", "%")
-    return query
-
-
-async def csv_to_parquet(
-    file_path: str,
-    inspection: dict,
-    resource_id: str | None = None,
-    check_id: int | None = None,
-    table_name: str | None = None,
-) -> tuple[str, int] | None:
-    """
-    Convert a csv file to parquet using inspection data.
-
-    If table_name is provided, reads directly from the PostgreSQL table
-    instead of re-reading and re-casting the CSV file.
-
-    Args:
-        file_path: CSV file path to convert.
-        inspection: CSV detective report.
-        resource_id: used to name the parquet file.
-        check_id: check ID to update with parquet info.
-        table_name: if provided, read from this PostgreSQL table instead of file_path.
-
-    Returns:
-        parquet_url: URL of the parquet file.
-        parquet_size: size of the parquet file.
-    """
-    if not config.CSV_TO_PARQUET and not config.DB_TO_PARQUET:
-        log.debug("CSV_TO_PARQUET and DB_TO_PARQUET turned off, skipping parquet export.")
-        return
-
-    if int(inspection.get("total_lines", 0)) < config.MIN_LINES_FOR_PARQUET:
-        log.debug(
-            f"Skipping parquet export for {resource_id} because it has less than {config.MIN_LINES_FOR_PARQUET} lines."
-        )
-        return
-
-    log.debug(
-        f"Converting from {engine_to_file.get(inspection.get('engine', ''), 'CSV')} "
-        f"to parquet for {resource_id} and sending to Minio."
-    )
-
-    if resource_id:
-        # Update resource status to CONVERTING_TO_PARQUET
-        await Resource.update(resource_id, {"status": "CONVERTING_TO_PARQUET"})
-
-    if config.DB_TO_PARQUET and table_name:
-        parquet_file, _ = await db_to_parquet(
-            table_name=table_name,
-            inspection=inspection,
-            output_filename=resource_id,
-        )
-    else:
-        parquet_file, _ = save_as_parquet(
-            records=generate_records(file_path, inspection, cast_json=False),
-            columns=inspection["columns"],
-            output_filename=resource_id,
-        )
-    parquet_size: int = os.path.getsize(parquet_file)
-    parquet_url: str = minio_client.send_file(parquet_file)
-
-    await Check.update(
-        check_id,
-        {
-            "parquet_url": parquet_url,
-            "parquet_size": parquet_size,
-        },
-    )
-    # returning only for tests purposes
-    return parquet_url, parquet_size
-
-
-async def csv_to_db(
-    file_path: str,
-    inspection: dict,
-    table_name: str,
-    table_indexes: dict[str, str] | None = None,
-    resource_id: str | None = None,
-    debug_insert: bool = False,
-) -> None:
-    """
-    Convert a csv file to database table using inspection data. It should (re)create one table:
-    - `table_name` with data from `file_path`
-
-    :df_chunks: chunks of pre-cast dataframes read from the file
-    :inspection: CSV detective report
-    :table_name: used to create tables
-    :debug_insert: insert record one by one instead of using postgresql COPY
-    """
-    if not config.CSV_TO_DB:
-        log.debug("CSV_TO_DB turned off, skipping.")
-        return
-
-    log.debug(
-        f"Converting from {engine_to_file.get(inspection.get('engine', ''), 'CSV')} "
-        f"to db for {table_name}"
-    )
-
-    if any(
-        sum(len(char.encode("utf-8")) for char in col) > config.NAMEDATALEN - 1
-        for col in inspection["columns"]
-    ):
-        raise ParseException(
-            step="scan_column_names",
-            resource_id=resource_id,
-            table_name=table_name,
-        ) from ValueError(
-            f"Column names cannot exceed {config.NAMEDATALEN - 1} characters in Postgres"
-        )
-
-    if resource_id:
-        # Update resource status to INSERTING_IN_DB
-        await Resource.update(resource_id, {"status": "INSERTING_IN_DB"})
-
-    # build a `column_name: type` mapping and explicitely rename reserved column names
-    columns = {db_col_name(c): helpers.get_python_type(v) for c, v in inspection["columns"].items()}
-
-    q = f'DROP TABLE IF EXISTS "{table_name}"'
-    db = await context.pool("csv")
-    await db.execute(q)
-
-    # Create table
-    q = compute_create_table_query(table_name=table_name, columns=columns, indexes=table_indexes)
-    try:
-        await db.execute(q)
-    except Exception as e:
-        raise ParseException(
-            message=str(e),
-            step="create_table_query",
-            resource_id=resource_id,
-            table_name=table_name,
-        ) from e
-
-    # this use postgresql COPY from an iterator, it's fast but might be difficult to debug
-    if not debug_insert:
-        # NB: also see copy_to_table for a file source
-        try:
-            await db.copy_records_to_table(
-                table_name,
-                records=generate_records(file_path, inspection, cast_json=False),
-                columns=list(columns.keys()),
-            )
-        except Exception as e:  # I know what I'm doing, pinky swear
-            raise ParseException(
-                message=str(e),
-                step="copy_records_to_table",
-                resource_id=resource_id,
-                table_name=table_name,
-            ) from e
-    # this inserts rows from iterator one by one, slow but useful for debugging
-    else:
-        bar = ProgressBar(total=inspection["total_lines"])
-        for r in bar.iter(generate_records(file_path, inspection, cast_json=False)):
-            data = {k: v for k, v in zip(inspection["columns"], r)}
-            # NB: possible sql injection here, but should not be used in prod
-            q = compute_insert_query(table_name=table_name, data=data, returning="__id")
-            await db.execute(q, *data.values())
-
-
-async def csv_to_db_index(
-    table_name: str, inspection: dict, check: Record | dict, dataset_id: str
-) -> None:
-    """Store meta info about a converted CSV table in `DATABASE_URL_CSV.tables_index`"""
-    db = await context.pool("csv")
-    q = "INSERT INTO tables_index(parsing_table, csv_detective, resource_id, dataset_id, url) VALUES($1, $2, $3, $4, $5)"
-    try:
-        await db.execute(
-            q,
-            table_name,
-            json.dumps(sanitize_for_json(inspection), default=str),
-            check.get("resource_id"),
-            dataset_id,
-            check.get("url"),
-        )
-    except Exception as e:
-        raise ParseException(
-            message=str(e),
-            step="tables_index_insertion",
-            resource_id=check.get("resource_id"),
-            url=check.get("url"),
-            check_id=check["id"],
-        ) from e
