+++ udata_hydra/analysis/tables_index.py
@@ -1,0 +1,71 @@
+import json
+import logging
+
+from asyncpg import Record
+from asyncpg.exceptions import UndefinedTableError
+from csv_detective.utils import sanitize_for_json
+
+from udata_hydra import context
+from udata_hydra.utils import ParseException
+
+log = logging.getLogger("udata-hydra")
+
+
+async def get_previous_analysis(resource_id: str) -> dict | None:
+    db = await context.pool("csv")
+    q = (
+        "SELECT parsing_table, csv_detective FROM tables_index "
+        f"WHERE resource_id='{resource_id}' ORDER BY created_at DESC LIMIT 1"
+    )
+    res = await db.fetch(q)
+    if not res:
+        return None
+    analysis = json.loads(res[0]["csv_detective"])
+    # the csv_detective column is JSONB, so keys are reordered compared to the actual table
+    # so we get the right order from the table
+    try:
+        rows: list[Record] = await db.fetch(f'SELECT * FROM "{res[0]["parsing_table"]}" LIMIT 1')
+    except UndefinedTableError:
+        log.error(f"Could not find table {res[0]['parsing_table']} for resource {resource_id}")
+        return None
+    if len(rows) == 0:
+        return None
+    # the __id column is generated by hydra, not natively in the data
+    table_columns = [col for col in rows[0].keys() if col != "__id"]
+    analysis["columns"] = {col: analysis["columns"][col] for col in table_columns}
+    return analysis
+
+
+async def csv_to_db_index(
+    table_name: str, inspection: dict, check: Record | dict, dataset_id: str
+) -> None:
+    """Store meta info about a converted CSV table in `DATABASE_URL_CSV.tables_index`"""
+    db = await context.pool("csv")
+    q = "INSERT INTO tables_index(parsing_table, csv_detective, resource_id, dataset_id, url) VALUES($1, $2, $3, $4, $5)"
+    try:
+        await db.execute(
+            q,
+            table_name,
+            json.dumps(sanitize_for_json(inspection), default=str),
+            check.get("resource_id"),
+            dataset_id,
+            check.get("url"),
+        )
+    except Exception as e:
+        raise ParseException(
+            message=str(e),
+            step="tables_index_insertion",
+            resource_id=check.get("resource_id"),
+            url=check.get("url"),
+            check_id=check["id"],
+        ) from e
+
+
+async def parquet_to_db_index(
+    table_name: str,
+    inspection: dict,
+    check: Record,
+    dataset_id: str,
+) -> None:
+    # convenience method just to make it clearer
+    await csv_to_db_index(table_name, inspection, check, dataset_id)
