+++ udata_hydra/cli.py
@@ -20,10 +20,12 @@ from progressist import ProgressBar
 
 from udata_hydra import config
 from udata_hydra.analysis.csv import analyse_csv, csv_detective_routine
-from udata_hydra.analysis.geojson import analyse_geojson, csv_to_geojson, geojson_to_pmtiles
+from udata_hydra.analysis.geojson import analyse_geojson
 from udata_hydra.analysis.ogc import analyse_ogc
 from udata_hydra.analysis.parquet import analyse_parquet
 from udata_hydra.analysis.resource import analyse_resource
+from udata_hydra.conversion.csv_to_geojson import csv_to_geojson
+from udata_hydra.conversion.geojson_to_pmtiles import geojson_to_pmtiles
 from udata_hydra.crawl.check_resources import (
     check_resource as crawl_check_resource,
 )
