+++ udata_hydra/conversion/__init__.py
@@ -1,0 +1,12 @@
+import csv as stdcsv
+import sys
+
+# Increase CSV field size limit to maximum possible
+# https://stackoverflow.com/a/15063941
+field_size_limit = sys.maxsize
+while True:
+    try:
+        stdcsv.field_size_limit(field_size_limit)
+        break
+    except OverflowError:
+        field_size_limit = int(field_size_limit / 10)
